+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* Added `volume` parameter to control video sound in AppBar.
+* Added option to enable audio with customizable unmute button on web.
+* Updated `video_player` dependency to latest version.
+
 ## 0.0.1+2
 
 * Updating example and readme